Move omitTypename to object folder
