+++ apps/spruce/src/pages/preferences/preferencesTabs/NotificationsTab.tsx
@@ -16,12 +16,10 @@ import {
 } from "gql/generated/types";
 import { UPDATE_USER_SETTINGS } from "gql/mutations";
 import { useUserSettings } from "hooks";
-import { string } from "utils";
 import { reportError } from "utils/errorReporting";
+import { omitTypename } from "utils/object";
 import { NotificationField } from "./notificationTab/NotificationField";
 import { UserSubscriptions } from "./notificationTab/UserSubscriptions";
-
-const { omitTypename } = string;
 
 export const NotificationsTab: React.FC = () => {
   const dispatchToast = useToastContext();
